+++ aldrin/CHANGELOG.md
@@ -9,12 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support protocol version 1.18.
 - Add `low_level::Service::type_id()` getter.
 - Add `low_level::Service::query_introspection()`.
 
 ### Fixed
 
 - Events are now only emitted to the broker if there are actual subscribers.
+- On protocol version 1.18, `low_level::Proxy::next_event()` and `poll_next_event()` now return
+  `None` when the service destroyed even if no events are subscribed.
 
 ### Changed